Inventory (WIP)
Currently working on sorting out an inventory system. Most of the work doesn't directly affect any of the already done stuff in the GameContent folder, but does draw from it.

Once the Inventory is in a workable condition, I'll move on to trying to sort out the Hotbar/Quick-Use Wheel.