Basic Inventory And Pickup System
Inventory has been made. Hard-coded to 10 slots currently. Can bring it up with the 'I' key on keyboard. Also makes the cursor visible so you can actually interact with the menu. There's an additional action menu that pops up when you click on an item in your inventory. Currently have three of the four actions there working, 'Use', 'Drop' and 'Cancel'.

The item template is somewhat finished. I've dropped one copy of it in to the game world to interact with. Basic cube for now but it's coded to register as 'Bread'. You can pick items up with the 'E' key. The template bread is currently made to bump up your health and stamina when used, which can be done via the inventory for now.

There is a small visual problem with the inventory where the slots seem to be compact unless you have something in one of the inventory slots. Not entirely sure what the issue there is, but probably something with settings that I'll try and work out when I can.